+++ iqm_client/src/iqm/qiskit_iqm/examples/bell_measure.py
@@ -11,8 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""This file is an example of using Qiskit on IQM to run a simple but non-trivial quantum circuit on an IQM quantum
-computer. See the Qiskit on IQM user guide for instructions:
+"""An example on using Qiskit on IQM to run a simple quantum circuit on an IQM quantum computer.
+
+See the Qiskit on IQM user guide for instructions:
 https://docs.meetiqm.com/iqm-client/user_guide_qiskit.html
 """
 
@@ -22,34 +23,58 @@ from iqm.qiskit_iqm.iqm_provider import IQMProvider
 from qiskit import QuantumCircuit, transpile
 
 
-def bell_measure(server_url: str) -> dict[str, int]:
-    """Run a circuit that prepares and measures a Bell state.
+def bell_measure(server_url: str, token: str | None = None, shots: int = 1000) -> dict[str, int]:
+    """Execute a quantum circuit that prepares and measures a generalized Bell (aka GHZ) state.
 
     Args:
         server_url: URL of the IQM server used for execution
+        token: API token for authentication. If not given, uses :env:`IQM_TOKEN`.
+        shots: Requested number of shots.
 
     Returns:
-        a mapping of bitstrings representing qubit measurement results to counts for each result
+        Mapping of bitstrings representing qubit measurement results to counts for each result.
 
     """
-    backend = IQMProvider(server_url).get_backend()
+    print(f"Executing a circuit on {server_url}")
+    # Initialize a backend without metrics as IQMClient._get_calibration_quality_metrics is not supported by resonance
+    backend = IQMProvider(server_url, token=token).get_backend()
     if backend.num_qubits < 2:
         raise ValueError("We need two qubits for the Bell state.")
-    circuit = QuantumCircuit(2)
-    circuit.h(0)
-    circuit.cx(0, 1)
-    circuit.measure_all()
 
-    new_circuit = transpile(circuit, backend)
-    return backend.run(new_circuit, shots=1000).result().get_counts()
+    # Just to make sure that "get_static_quantum_architecture" method works
+    static_quantum_architecture = backend.client.get_static_quantum_architecture()
+    print(f"static_quantum_architecture={static_quantum_architecture}")
+
+    # Define a quantum circuit for a GHZ state
+    n_qubits = min(backend.num_qubits, 5)  # use at most 5 qubits
+    qc = QuantumCircuit(n_qubits)
+    qc.h(0)
+    for qb in range(1, n_qubits):
+        qc.cx(0, qb)
+    qc.barrier()
+    qc.measure_all()
+
+    # Transpile the circuit
+    qc_transpiled = transpile(qc, backend)
+    print(qc_transpiled.draw(output="text"))
+
+    # Run the circuit
+    job = backend.run(qc_transpiled, shots=shots)
+    return job.result().get_counts()
 
 
 if __name__ == "__main__":
     argparser = argparse.ArgumentParser()
     argparser.add_argument(
-        "--url",
-        help="IQM server URL",
-        # For example https://cocos.resonance.meetiqm.com/garnet
-        default="https://<IQM SERVER>",
+        "--url", required=True, help='IQM server URL, for example "https://cocos.resonance.meetiqm.com/garnet"'
     )
-    print(bell_measure(argparser.parse_args().url))
+    argparser.add_argument(
+        "--token",
+        help="API token for authentication",
+        # Provide the API token explicitly or set it as an environment variable
+        # following the https://docs.meetiqm.com/iqm-client/user_guide_qiskit.html#authentication
+    )
+
+    args = argparser.parse_args()
+    counts = bell_measure(args.url, args.token)
+    print(counts)